health(criteria): #78 moved specs to source folder + rename + regroup
