+++ src/criteria/test/rendering.spec.ts
@@ -1,6 +1,60 @@
-import { inRange } from "../../../src";
+import { inRange, inSet, notInSet, or } from "../value-criterion";
 
-describe("renderFromToValueCriterion", () => {
+describe("render: in-set", () => {
+    it("should render correctly", () => {
+        // arrange
+        const criterion = inSet([1, 2, 3]);
+        const expected = "{1, 2, 3}";
+
+        // act
+        const actual = criterion.toString();
+
+        // assert
+        expect(actual).toEqual(expected);
+    });
+});
+
+describe("render: not-in-set", () => {
+    it("should render correctly", () => {
+        // arrange
+        const criterion = notInSet([1, 2, 3]);
+        const expected = "!{1, 2, 3}";
+
+        // act
+        const actual = criterion.toString();
+
+        // assert
+        expect(actual).toEqual(expected);
+    });
+});
+
+describe("render: value-criteria", () => {
+    it("should render brackets even if there is only 1 element", () => {
+        // arrange
+        const criteria = or([inSet([1, 2, 3])]);
+        const expected = "({1, 2, 3})";
+
+        // act
+        const rendered = criteria.toString();
+
+        // assert
+        expect(rendered).toEqual(expected);
+    });
+
+    it("should render correctly", () => {
+        // arrange
+        const criteria = or([inSet([1, 2, 3]), inRange(0, 7)]);
+        const expected = "({1, 2, 3} | [0, 7])";
+
+        // act
+        const rendered = criteria.toString();
+
+        // assert
+        expect(rendered).toEqual(expected);
+    });
+});
+
+describe("render: in-range", () => {
     it("[1, 7] should be rendered correctly", () => {
         // arrange
         const criterion = inRange(1, 7);
