--- test/unit/criteria/render-not-in-value-criterion.spec.ts
@@ -1,15 +1,0 @@
-import { notInSet } from "../../../src";
-
-describe("renderNotInValueCriterion", () => {
-    it("should render correctly", () => {
-        // arrange
-        const criterion = notInSet([1, 2, 3]);
-        const expected = "!{1, 2, 3}";
-
-        // act
-        const actual = criterion.toString();
-
-        // assert
-        expect(actual).toEqual(expected);
-    });
-});
